requirements.txt
pywin32
pyyaml
python-docx
python-pptx


